+++ src/main/java/com/springmvc/website/services/PostService.java
@@ -8,6 +8,6 @@ public interface PostService {
 	public void addPost(Post post);
 	public void updatePost(Post post);
 	public List<Post> ListPosts();
-	public Post getPostById(Long id);
-	public void removePost(Long id);
+	public Post getPostById(int id);
+	public void removePost(int id);
 }
